+++ src/interface/constants.ts
@@ -4,3 +4,10 @@ export const IMAGE_URL =
 export const API_URL_PEOPLE = 'https://swapi.dev/api/people';
 
 export const ITEMS_PER_PAGE_INITIAL: number = 10;
+
+export enum ItemsPerPage {
+  Five = 5,
+  Ten = 10,
+  TwentyFive = 25,
+  Forty = 40,
+}
