+++ tsconfig.json
@@ -18,7 +18,7 @@
     "allowJs": false,
     "baseUrl": ".",
     "paths": {
-      "src:/*":["src:/*"],
+      "src:/*": ["src:/*"]
     },
 
     /* Linting */
